+++ examples/oft-solana/tasks/solana/initConfig.ts
@@ -1,4 +1,4 @@
-import { Keypair, PublicKey } from '@solana/web3.js'
+import { PublicKey } from '@solana/web3.js'
 import { ConfigurableTaskDefinition } from 'hardhat/types'
 
 import { inheritTask } from '@layerzerolabs/devtools-evm-hardhat'
@@ -18,7 +18,6 @@ const wireLikeTask = inheritTask(TASK_LZ_OAPP_WIRE)
 interface Args {
     logLevel: LogLevel
     solanaEid: EndpointId
-    solanaSecretKey?: Keypair
     multisigKey?: PublicKey
     internalConfigurator?: OAppConfigurator
 }