+++ packages/protocol-devtools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @layerzerolabs/protocol-devtools
 
+## 1.1.5
+
+### Patch Changes
+
+- e256387: Updating packages
+- Updated dependencies [e256387]
+  - @layerzerolabs/devtools@0.4.6
+  - @layerzerolabs/io-devtools@0.1.16
+
 ## 1.1.4
 
 ### Patch Changes